- Lint - Move PluginInstance to /chat
