+++ src/twitch/caster.mts
@@ -1,7 +1,6 @@
 import { TWITCH_ENVIRONMENT } from "../environment.mts";
 import { TwitchOIDC } from "./oidc.mjs";
 import type { TwitchIrcClient } from "./irc.mts";
-import type { websocketServer } from "../http/websockets.mts";
 import EventEmitter from "events";
 import { WebSocket as WS } from "ws";
 import { Logger } from "../logging.mjs";
@@ -9,7 +8,6 @@ import { serializeError } from "serialize-error";
 import VError from "verror";
 import type {
   EventsubMessage,
-  EventsubNotificationMessage,
   EventsubWelcomeMessage,
 } from "./api/eventsub.mjs";
 
@@ -23,7 +21,7 @@ export class TwitchCasterClient extends EventEmitter {
   constructor(
     private oidc: TwitchOIDC | null = null,
     public readonly irc: TwitchIrcClient,
-    private readonly server: ReturnType<typeof websocketServer>,
+    // private readonly server: ReturnType<typeof websocketServer>
   ) {
     super();
   }
@@ -84,13 +82,13 @@ export class TwitchCasterClient extends EventEmitter {
                 textData,
               },
             },
-            textData,
+            textData
           );
         }
       }
 
       const data = JSON.parse(
-        event.data.toString(),
+        event.data.toString()
       ) as unknown as EventsubMessage;
 
       switch (data.metadata.message_type) {
@@ -121,7 +119,7 @@ export class TwitchCasterClient extends EventEmitter {
         case "session_keepalive":
           break;
         case "notification":
-          const notification = data as EventsubNotificationMessage;
+          // const notification = data as EventsubNotificationMessage;
 
           // const parsedSubscription = {
           //   ...notification.payload,
